# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: F1 Strategy Intelligence Engine API
  version: 0.1.0
  description: API contract scaffold for the F1 strategy platform.
servers:
  - url: http://localhost:8000
paths:
...
  /api/v1/predictions/undercut:
    post:
      summary: Request undercut prediction
      responses:
        "202":
          description: Prediction accepted
        "501":
          description: Undercut model inference is not implemented in the scaffold
  /api/v1/predictions/overcut:
    post:
      summary: Request overcut prediction
      responses:
        "202":
          description: Prediction accepted
        "501":
          description: Overcut model inference is not implemented in the scaffold
  /api/v1/predictions/tyres:
    post:
      summary: Request tyre prediction
      responses:
        "202":
          description: Prediction accepted
        "501":
          description: Tyre model inference is not implemented in the scaffold
...